+++ packages/core/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The `AbstractVideoIngestionAppliance` now only internally tracks the position of the video stream in a given MPEG-TS stream, as opposed to tracking all streams.
+
 ### Fixed
 - The optional `baseTime` parameter is now in terms of the fraction instead of the denominator of the fraction (e.g. `1/90000` instead of `90000`).
 - `AbstractVideoIngestionAppliance` now supports streams that have been running long enough for a PTS rollover to occur.